[instancer] Split featureVars code into own module
Towards fixing https://github.com/fonttools/fonttools/issues/2737
